Move auth classes in their own directory
