+++ lib/auth/auth.js
