+++ lib/auth/github_auth.js
@@ -4,7 +4,7 @@ var fs = require('fs');
 var prompt = require('prompt-promise');
 var Request = require('superagent');
 
-var credentials = require('../github_credentials.json');
+var credentials = require('../../github_credentials.json');
 
 var Auth = require('./auth');
