+++ lib/auth/gitlab_auth.js
