+++ lib/auth/google_auth.js
@@ -4,7 +4,7 @@ var googleAuth = require('google-auth-library');
 var Promise = require('bluebird');
 var prompt = require('prompt-promise');
 
-var client_secret = require('../client_secret.json');
+var client_secret = require('../../client_secret.json');
 
 var Auth = require('./auth');
