+++ lib/auth/jira_auth.js
@@ -15,8 +15,8 @@ var Auth = require('./auth');
 // To generate a new key pair, the following commands were used (based on https://www.madboa.com/geek/openssl/#key-rsa)
 // openssl genrsa -out jira-rsa-key.pem 2048 # generate a new key
 // openssl rsa -in jira-rsa-key.pem -pubout # print public key
-var privateKeyData = fs.readFileSync(__dirname + '/../jira-rsa-key.pem', "utf8");
-var credentials = require('../jira_credentials.json');
+var privateKeyData = fs.readFileSync(__dirname + '/../../jira-rsa-key.pem', "utf8");
+var credentials = require('../../jira_credentials.json');
 
 class JiraAuth extends Auth {
     constructor() {
