+++ lib/auth/null_auth.js
