+++ lib/auth/slack_auth.js
@@ -5,7 +5,7 @@ var readline = require('readline');
 var prompt = require('prompt-promise');
 var Slack = require('slack-api').promisify();
 
-var credentials = require('../slack_credentials.json');
+var credentials = require('../../slack_credentials.json');
 
 var Auth = require('./auth');
