+++ lib/auth/zebra_auth.js
